+++ app/actions/index.ts
@@ -16,6 +16,7 @@ import fileImport from './fileImport';
 import git from './git';
 import keys from './keys';
 import privateKey from './privateKey';
+import './recentPrivateKeys';
 import './recentRepositories';
 import repository from './repository';
 import search from './search';